confluence-mcp/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "confluence-mcp"
version = "0.1.0"
description = "MCP server for Atlassian Confluence - read, write, update pages with multi-tabs and tables support"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Developer", email = "developer@example.com"}
]
keywords = ["mcp", "confluence", "atlassian", "llm", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "markdownify>=0.11.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-httpx>=0.21.0",
]

[project.scripts]
confluence-mcp = "confluence_mcp.server:main"

[project.urls]
Homepage = "https://github.com/example/confluence-mcp"
Repository = "https://github.com/example/confluence-mcp"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
